+++ src/frontend/apps/e2e/__tests__/app-impress/doc-grid-dnd.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { createDoc, mockedListDocs } from './common';
+import { createDoc, mockedListDocs } from './utils-common';
 
 test.describe('Doc grid dnd', () => {
   test('it creates a doc', async ({ page, browserName }) => {
