+++ src/frontend/apps/e2e/__tests__/app-impress/doc-version.spec.ts
@@ -5,7 +5,7 @@ import {
   goToGridDoc,
   mockedDocument,
   verifyDocName,
-} from './common';
+} from './utils-common';
 
 test.beforeEach(async ({ page }) => {
   await page.goto('/');
