+++ src/frontend/apps/e2e/__tests__/app-impress/home.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { overrideConfig } from './common';
+import { overrideConfig } from './utils-common';
 
 test.beforeEach(async ({ page }) => {
   await page.goto('/docs/');
